+++ rework_pysatl_mpest/initializers/strategies.py
